A lot of update get get the package ready for publishing and use for benchmark and monitoring training only

+++ dl_classification_model/processing/utils.py
@@ -3,6 +3,8 @@ import joblib
 from sklearn.metrics import accuracy_score
 from sklearn.metrics import classification_report
 import logging
+import tensorflow as tf
+from sklearn.pipeline import Pipeline
 
 from dl_classification_model.config import config
 from dl_classification_model import __version__ as _version
@@ -29,7 +31,7 @@ def load_dataset(filename):
     return _data
 
 
-def save_pipeline(pipeline_to_save,save_file_name=f"{config.MODEL_PIPELINE_NAME}{_version}.pkl"):
+def save_pipeline(pipeline_to_save,save_file_name=f"{config.MODEL_PIPELINE_NAME}{_version}",dataprep_name = f"{config.DATAPREP_PIPELINE_NAME}{_version}.pkl"):
     """ Save a pipeline 
     Save a pipeline and overwrites a pipeline of the same version as this one.
 
@@ -41,16 +43,29 @@ def save_pipeline(pipeline_to_save,save_file_name=f"{config.MODEL_PIPELINE_NAME}
     
     # setting up the versioned filename
     #save_file_name = f"{config.MODEL_PIPELINE_NAME}{_version}.pkl"
-    save_file_path = config.TRAINED_MODEL_DIR/save_file_name
+    dataprep_save_file_path = config.TRAINED_MODEL_DIR/dataprep_name
+    model_save_file_path = config.TRAINED_MODEL_DIR/save_file_name
+
+
+    # saving the pipeline
+    dataprep = pipeline_to_save.steps[:-1]
+    classes_name_file = f"{save_file_name}_classes_names_v{_version}.pkl"
+    joblib.dump(pipeline_to_save.named_steps['model'].classes_,f"{config.TRAINED_MODEL_DIR / classes_name_file}")
+    
+    joblib.dump(dataprep,dataprep_save_file_path)
+
+    pipeline_to_save.named_steps['model'].model.save(str(model_save_file_path))
 
     # remove older pipelines
-    remove_old_pipelines(files_to_keep=[save_file_name])
-    # saving the pipeline
-    joblib.dump(pipeline_to_save,save_file_path)
-    # logging
-    _logger.info(f"The pipeline was saved as : {save_file_name}")
+    remove_old_pipelines(files_to_keep=[save_file_name,dataprep_name,classes_name_file])
 
-def load_pipeline(file_name):
+    # logging
+    _logger.info(f"The dataprep pipeline was saved as : {dataprep_name}")
+    _logger.info(f"The model was saved as : {save_file_name}")
+
+    _logger.info(f"The model classes names were saved as : {config.TRAINED_MODEL_DIR / classes_name_file}")
+
+def load_pipeline(file_name = {"dataprep_name": f"{config.DATAPREP_PIPELINE_NAME}{_version}.pkl","model_name" : f"{config.MODEL_PIPELINE_NAME}{_version}"}):
     """ load a saved pipeline 
     
     Parameters
@@ -64,11 +79,33 @@ def load_pipeline(file_name):
         A fitted pipeline.
     """
 
-    file_path = config.TRAINED_MODEL_DIR/file_name
-    _fitted_pipeline = joblib.load(file_path)
-    _logger.info(f"The pipeline {file_name} has been loaded")
-    return _fitted_pipeline
+    data_prep_file_path = config.TRAINED_MODEL_DIR/file_name["dataprep_name"]
+    model_file_path = config.TRAINED_MODEL_DIR/file_name["model_name"]
 
+    _data_prep_fitted_pipeline = joblib.load(data_prep_file_path)
+    _logger.info(f"The pipeline {data_prep_file_path} has been loaded")
+
+    _logger.info(f"The model {model_file_path} has been loaded")
+
+    build_model = lambda: tf.keras.models.load_model(model_file_path)
+
+    classifier = tf.keras.wrappers.scikit_learn.KerasClassifier(build_fn=build_model)
+    classes_name_file = f"{config.MODEL_PIPELINE_NAME}{_version}_classes_names_v{_version}.pkl"
+
+    classifier.classes_ = joblib.load(f"{config.TRAINED_MODEL_DIR / classes_name_file}")
+    classifier.model = build_model()
+
+    return Pipeline(_data_prep_fitted_pipeline+[("model",classifier)])
+    
+def rm_tree(file_path):
+    """ Remove file and folder with pathlib"""
+    if file_path.is_file():
+        file_path.unlink()
+    else:
+        for child in file_path.iterdir():
+            rm_tree(child)
+
+        file_path.rmdir()
 
 def remove_old_pipelines(files_to_keep):
     """  Remove previous pipeline based on the version.
@@ -82,13 +119,21 @@ def remove_old_pipelines(files_to_keep):
     """
 
     # adding __init__.py to the not delete list as we do not want to delete it.
-    to_not_remove = files_to_keep + ["__init__.py"]
+    model_file_name=f"{config.MODEL_PIPELINE_NAME}{_version}"
+    dataprep_name = f"{config.DATAPREP_PIPELINE_NAME}{_version}.pkl"
+    best_params_file = f"best_params_v{_version}.pkl"
+    classes_name_file = f"{config.MODEL_PIPELINE_NAME}{_version}_classes_names_v{_version}.pkl"
+
+    to_not_remove = files_to_keep + ["__init__.py",best_params_file,model_file_name,dataprep_name,classes_name_file]
     removed_files = []
     for file in config.TRAINED_MODEL_DIR.iterdir():
         if file.name not in to_not_remove:
-            file.unlink()
-            removed_files.append(file.name)
-    
+            if file.is_file():
+                file.unlink()
+                removed_files.append(file.name)
+            else:
+                rm_tree(file)
+                removed_files.append(file.name)
     # Logging the removal of files if there was.
     if len(removed_files)>0:
         _logger.info(f"file/pipelines removed : {removed_files}")
